watchdog: pic32-dmt: PIC32 deadman timer driver support.
The primary function of the deadman timer (DMT) is to reset the processor
in the event of a software malfunction. The DMT is a free-running
instruction fetch timer, which is clocked whenever an instruction fetch
occurs until a count match occurs. Instructions are not fetched when
the processor is in Sleep mode.

A DMT is typically used in mission critical and safety critical applications,
where any single failure of the software functionality and sequencing must be
detected.

Signed-off-by: Purna Chandra Mandal <purna.mandal@microchip.com>