MIPS: pic32: Fix incomplete earlyprintk= command line parsing.
This fixes parsing the command line earlyprintk= arguments only in some cases.
The port and baud rate can be specified as arguments and they were not being
parsed at all as intended. To support passing argument through the DTB,
initialization of the early printk had to be moved until after that is parsed.

The only early console ports supported are ttyS1 and ttyS5.  The default is
ttyS1,115200.

Signed-off-by: Joshua Henderson <joshua.henderson@microchip.com>