usb: musb: Fix DMA for host mode
Fix done condition that advances the musb schedule. Without this patch the the
msub_advance_schedule() is called immediately after receiving an endpoint rx
interrupt without waiting for the dma transfer to complete. As a consequence
when the dma complete interrupt arrives the in_qh member of hw_ep is already
null an the musb_host_rx() exits on !urb error case.

Signed-off-by: Cristian Birsan <cristian.birsan@microchip.com>