MIPS: (opt) Fix of reading I-pages from non-DMA FS devices for ID cache separation
This optional fix provides a D-cache flush for instruction code pages on
page faults. In case of non-DMA block device a driver doesn't know that it
reads I-page and doesn't flush D-cache generally on systems without
cache aliasing. And that takes toll during page fault of instruction pages.

It is not a perfect fix, it should be considered as a temporary fix.
The permanent fix would track page origin in page cache and flushes
D-cache during reception of page from driver only but not at each page fault.
It is not done yet.

Author: Leonid Yegoshin <yegoshin@mips.com>
Signed-off-by: Leonid Yegoshin <yegoshin@mips.com>

http://patchwork.linux-mips.org/patch/8635/